新增model将库表对应实体类与DTO、VO分离，新增超级管理菜单权限

+++ SAND-Bussiness/ls-sys/sys-provider/src/main/java/com/sand/sys/service/impl/SysRoleServiceImpl.java
@@ -17,6 +17,7 @@ import com.sand.sys.entity.SysRole;
 import com.sand.sys.entity.SysRoleMenu;
 import com.sand.sys.enums.RoleEnum;
 import com.sand.sys.mapper.SysRoleMapper;
+import com.sand.sys.model.SysRoleModel;
 import com.sand.sys.service.ISysRoleMenuService;
 import com.sand.sys.service.ISysRoleService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,29 +41,29 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
 
   @Override
   @Transactional(rollbackFor = LsException.class)
-  public int add(SysRole role) {
+  public int add(SysRoleModel model) {
     // 参数校验
-    checkedSysRole(role, OperateEnum.INSERT);
+    checkedSysRole(model, OperateEnum.INSERT);
     // 信息入库
-    if (!super.save(role)) {
+    if (!super.save(model)) {
       throw new LsException("新增角色信息入库异常！");
     }
     // 重新授权
-    reauthorize(role);
+    reauthorize(model);
     return 0;
   }
 
   @Override
   @Transactional(rollbackFor = LsException.class)
-  public int edit(SysRole role) {
+  public int edit(SysRoleModel model) {
     // 参数校验
-    checkedSysRole(role, OperateEnum.UPDATE);
+    checkedSysRole(model, OperateEnum.UPDATE);
     // 信息入库
-    if (!super.updateById(role)) {
+    if (!super.updateById(model)) {
       throw new LsException("修改角色信息入库异常！");
     }
     // 重新授权
-    reauthorize(role);
+    reauthorize(model);
     return 0;
   }
 
@@ -73,6 +74,10 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
     if (StringUtil.isBlank(roleId)) {
       throw new LsException(ResultEnum.PARAM_MISSING_ERROR, "角色ID未分配！");
     }
+    SysRole dbRole = super.getById(roleId);
+    if (Objects.isNull(dbRole)) {
+      throw new LsException(ResultEnum.PARAM_CHECKED_ERROR, "角色信息不存在！");
+    }
     QueryWrapper<SysRoleMenu> roleMenuWrapper = new QueryWrapper<>();
     roleMenuWrapper.eq("role_id", roleId);
     if (!roleMenuService.remove(roleMenuWrapper)) {
@@ -82,13 +87,13 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
 
   @Override
   @Transactional(rollbackFor = LsException.class)
-  public void reauthorize(SysRole role) {
+  public void reauthorize(SysRoleModel model) {
     // 取消旧的权限
-    this.cancelAuthorize(role.getRoleId());
+    this.cancelAuthorize(model.getRoleId());
     // 分配新的权限
     List<SysRoleMenu> roleMenuList = new ArrayList<>();
-    role.getMenuIds().forEach(menuId -> {
-      SysRoleMenu roleMenu = SysRoleMenu.builder().roleId(role.getRoleId()).menuId(menuId).build();
+    model.getMenuIds().forEach(menuId -> {
+      SysRoleMenu roleMenu = SysRoleMenu.builder().roleId(model.getRoleId()).menuId(menuId).build();
       roleMenuList.add(roleMenu);
     });
     if (!roleMenuService.saveBatch(roleMenuList)) {
@@ -99,56 +104,56 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
   /**
    * 参数校验
    *
-   * @param role dto
+   * @param model dto
    */
-  private void checkedSysRole(SysRole role, OperateEnum operate) {
+  private void checkedSysRole(SysRoleModel model, OperateEnum operate) {
     // 新增/修改通用参数非空校验
-    if (StringUtil.isBlank(role.getRoleName())) {
+    if (StringUtil.isBlank(model.getRoleName())) {
       throw new LsException(ResultEnum.PARAM_MISSING_ERROR, "角色名称不能为空！");
     }
-    if (StringUtil.isBlank(role.getRoleKey())) {
+    if (StringUtil.isBlank(model.getRoleKey())) {
       throw new LsException(ResultEnum.PARAM_MISSING_ERROR, "权限字符不能为空！");
     }
     // 校验角色状态是否存在
-    if (Objects.nonNull(role.getStatus())) {
-      RoleEnum.Status status = RoleEnum.Status.getByStatus(role.getStatus());
+    if (Objects.nonNull(model.getStatus())) {
+      RoleEnum.Status status = RoleEnum.Status.getByStatus(model.getStatus());
       if (Objects.isNull(status)) {
         throw new LsException(ResultEnum.PARAM_CHECKED_ERROR, "此角色状态不存在！");
       }
     }
     // 校验删除标志是否存在
-    if (Objects.nonNull(role.getDelFlag())) {
-      RoleEnum.DelFlag delFlag = RoleEnum.DelFlag.getByFlag(role.getDelFlag());
+    if (Objects.nonNull(model.getDelFlag())) {
+      RoleEnum.DelFlag delFlag = RoleEnum.DelFlag.getByFlag(model.getDelFlag());
       if (Objects.isNull(delFlag)) {
         throw new LsException(ResultEnum.PARAM_CHECKED_ERROR, "此删除标志不存在！");
       }
     }
     // 校验数据范围是否存在
-    if (Objects.nonNull(role.getDataScope())) {
-      RoleEnum.DataScope dataScope = RoleEnum.DataScope.getByScope(role.getDataScope());
+    if (Objects.nonNull(model.getDataScope())) {
+      RoleEnum.DataScope dataScope = RoleEnum.DataScope.getByScope(model.getDataScope());
       if (Objects.isNull(dataScope)) {
         throw new LsException(ResultEnum.PARAM_CHECKED_ERROR, "此数据范围不存在！");
       }
     }
     // 菜单名查询条件组装
     QueryWrapper<SysRole> roleNameWrapper = new QueryWrapper<>();
-    roleNameWrapper.eq("role_name", role.getRoleName());
+    roleNameWrapper.eq("role_name", model.getRoleName());
     if (Objects.equals(operate, OperateEnum.INSERT)) {
-      if (StringUtil.isNotBlank(role.getRoleId())) {
-        SysRole dbRole = super.getById(role.getRoleId());
+      if (StringUtil.isNotBlank(model.getRoleId())) {
+        SysRole dbRole = super.getById(model.getRoleId());
         if (Objects.nonNull(dbRole)) {
           throw new LsException(ResultEnum.PARAM_CHECKED_ERROR, "此角色信息已存在！");
         }
       }
     } else if (Objects.equals(operate, OperateEnum.UPDATE)) {
-      if (StringUtil.isBlank(role.getRoleId())) {
+      if (StringUtil.isBlank(model.getRoleId())) {
         throw new LsException(ResultEnum.PARAM_MISSING_ERROR, "角色ID不能为空！");
       }
-      SysRole dbRole = super.getById(role.getRoleId());
+      SysRole dbRole = super.getById(model.getRoleId());
       if (Objects.isNull(dbRole)) {
         throw new LsException(ResultEnum.PARAM_CHECKED_ERROR, "此角色信息不存在！");
       }
-      roleNameWrapper.ne("role_id", role.getRoleId());
+      roleNameWrapper.ne("role_id", model.getRoleId());
     }
     // 校验角色名称是否重复
     List<SysRole> roleNameList = super.list(roleNameWrapper);
